requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
praw==7.7.1
sentence-transformers==2.3.1
google-generativeai==0.3.2
groq==0.4.2
python-dotenv==1.0.1
sqlalchemy==2.0.25
requests==2.31.0
pydantic==2.6.0
pydantic-settings==2.1.0
apscheduler==3.10.4
jinja2==3.1.3
python-multipart==0.0.9
aiofiles==23.2.1
